# backend/internal/apiserver/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/audit:
        get:
            tags:
                - AuditService
            description: "ListAuditRecords 分页查询审计日志（category/subject/object/action/result/ip 过滤）。\r\n   REST: GET /v1/audit"
            operationId: AuditService_ListAuditRecords
            parameters:
                - name: category
                  in: query
                  schema:
                    type: string
                - name: subject
                  in: query
                  schema:
                    type: string
                - name: object
                  in: query
                  schema:
                    type: string
                - name: action
                  in: query
                  schema:
                    type: string
                - name: result
                  in: query
                  schema:
                    type: string
                - name: ip
                  in: query
                  schema:
                    type: string
                - name: page_size
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: page_token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListAuditRecordsResponse'
    /v1/audit/{id}:
        get:
            tags:
                - AuditService
            description: "GetAuditRecord 取单条审计日志详情。\r\n   REST: GET /v1/audit/{id}"
            operationId: AuditService_GetAuditRecord
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetAuditRecordResponse'
...
    /v1/file:
        get:
            tags:
                - FileService
            description: "ListFiles 列出文件（租户过滤自动；原始文件名/内容类型过滤）。\r\n   REST: GET /v1/file"
            operationId: FileService_ListFiles
            parameters:
                - name: file_name
                  in: query
                  schema:
                    type: string
                - name: mime_type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListFilesResponse'
    /v1/file/upload:
        post:
            tags:
                - FileService
            description: "UploadFile 上传文件：校验后写入 MinIO 并登记元数据。\r\n   REST(gateway): POST /v1/file/upload（JSON body，content 为 base64）\r\n   REST(gin):     POST /v1/file/upload（multipart form：file + directory）"
            operationId: FileService_UploadFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UploadFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UploadFileResponse'
    /v1/file/{id}:
        get:
            tags:
                - FileService
            description: "GetFile 取文件元数据。\r\n   REST: GET /v1/file/{id}"
            operationId: FileService_GetFile
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetFileResponse'
        delete:
            tags:
                - FileService
            description: "DeleteFile 删除文件（先删对象，再删元数据）。\r\n   REST: DELETE /v1/file/{id}"
            operationId: FileService_DeleteFile
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteFileResponse'
    /v1/file/{id}/download:
        get:
            tags:
                - FileService
            description: "DownloadFile 下载文件内容（bytes）。\r\n   REST(gateway): GET /v1/file/{id}/download（JSON）\r\n   REST(gin):     GET /v1/file/{id}/download（流式，Content-Disposition）"
            operationId: FileService_DownloadFile
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DownloadFileResponse'
    /v1/menu:
        get:
            tags:
                - MenuService
            description: |-
                ListMenus 列出菜单树（根节点数组，children 递归嵌套；孤儿节点跳过——
                 源项目 pagination.BuildTree 同语义）。
                   REST: GET /v1/menu
            operationId: MenuService_ListMenus
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMenusResponse'
        post:
            tags:
                - MenuService
            description: |-
                CreateMenu 创建菜单节点（parent_id 空 = 根节点；源 parent_id=0 语义）。
                   REST: POST /v1/menu
            operationId: MenuService_CreateMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateMenuResponse'
    /v1/menu/{id}:
        get:
            tags:
                - MenuService
            description: |-
                GetMenu 取菜单节点。
                   REST: GET /v1/menu/{id}
                 单数路径段与 DefaultGRPCObject("MenuService/...")="menu" 同源（P9 策略单写）。
            operationId: MenuService_GetMenu
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMenuResponse'
        put:
            tags:
                - MenuService
            description: |-
                UpdateMenu 更新菜单节点（空值/UNSPECIFIED 字段不改）。
                   REST: PUT /v1/menu/{id}
            operationId: MenuService_UpdateMenu
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateMenuResponse'
        delete:
            tags:
                - MenuService
            description: |-
                DeleteMenu 删除菜单节点（递归级联删除子树——源 QueryAllChildrenIds 语义）。
                   REST: DELETE /v1/menu/{id}
            operationId: MenuService_DeleteMenu
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteMenuResponse'
...
components:
    schemas:
        AuditRecord:
            type: object
            properties:
                id:
                    type: string
                tenant_id:
                    type: string
                category:
                    type: string
                subject:
                    type: string
                object:
                    type: string
                action:
                    type: string
                result:
                    type: string
                error:
                    type: string
                ip_address:
                    type: string
                user_agent:
                    type: string
                request_id:
                    type: string
                trace_id:
                    type: string
                time:
                    type: string
                    format: date-time
            description: "AuditRecord 审计日志条目。字段对齐 model.AuditRecord（源\r\n OperationAuditLog/LoginAuditLog 共有核心字段精简；\r\n Success 由 Result 推导，LogHash/Signature 合规链后续迭代）。"
...
        CreateMenuRequest:
            type: object
            properties:
                id:
                    type: string
                parent_id:
                    type: string
                type:
                    enum:
                        - TYPE_UNSPECIFIED
                        - TYPE_CATALOG
                        - TYPE_MENU
                        - TYPE_BUTTON
                    type: string
                    format: enum
                name:
                    type: string
                path:
                    type: string
                component:
                    type: string
                title:
                    type: string
                icon:
                    type: string
                order:
                    type: integer
                    format: int32
                remark:
                    type: string
        CreateMenuResponse:
            type: object
            properties:
                menu:
                    $ref: '#/components/schemas/Menu'
...
        DeleteFileResponse:
            type: object
            properties:
                deleted:
                    type: string
        DeleteMenuResponse:
            type: object
            properties:
                deleted:
                    type: string
...
        DownloadFileResponse:
            type: object
            properties:
                content:
                    type: string
                    format: bytes
                file_name:
                    type: string
                mime_type:
                    type: string
        File:
            type: object
            properties:
                id:
                    type: string
                provider:
                    type: string
                bucket_name:
                    type: string
                save_file_name:
                    type: string
                file_directory:
                    type: string
                file_name:
                    type: string
                extension:
                    type: string
                content_hash:
                    type: string
                size:
                    type: integer
                    format: uint32
                link_url:
                    type: string
                mime_type:
                    type: string
                created_by:
                    type: string
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
            description: File 文件元数据。字段对齐源 storage/service/v1 的 File（recordFile）。
        GetAuditRecordResponse:
            type: object
            properties:
                record:
                    $ref: '#/components/schemas/AuditRecord'
...
        GetFileResponse:
            type: object
            properties:
                file:
                    $ref: '#/components/schemas/File'
        GetMenuResponse:
            type: object
            properties:
                menu:
                    $ref: '#/components/schemas/Menu'
...
        ListAuditRecordsResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/AuditRecord'
                total:
                    type: integer
                    format: uint32
                next_page_token:
                    type: string
...
        ListFilesResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/File'
                total:
                    type: integer
                    format: uint32
        ListMenusResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Menu'
                total:
                    type: integer
                    description: 全部节点数（含子孙，非根数）；uint32 避免 proto3 JSON int64 字符串化。
                    format: uint32
...
        Menu:
            type: object
            properties:
                id:
                    type: string
                parent_id:
                    type: string
                type:
                    enum:
                        - TYPE_UNSPECIFIED
                        - TYPE_CATALOG
                        - TYPE_MENU
                        - TYPE_BUTTON
                    type: string
                    format: enum
                name:
                    type: string
                path:
                    type: string
                component:
                    type: string
                title:
                    type: string
                icon:
                    type: string
                order:
                    type: integer
                    format: int32
                status:
                    enum:
                        - STATUS_UNSPECIFIED
                        - STATUS_ON
                        - STATUS_OFF
                    type: string
                    format: enum
                remark:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Menu'
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
            description: Menu 菜单节点（自引用树）。
...
        UpdateMenuRequest:
            type: object
            properties:
                id:
                    type: string
                parent_id:
                    type: string
                type:
                    enum:
                        - TYPE_UNSPECIFIED
                        - TYPE_CATALOG
                        - TYPE_MENU
                        - TYPE_BUTTON
                    type: string
                    format: enum
                name:
                    type: string
                path:
                    type: string
                component:
                    type: string
                title:
                    type: string
                icon:
                    type: string
                order:
                    type: integer
                    format: int32
                order_set:
                    type: boolean
                status:
                    enum:
                        - STATUS_UNSPECIFIED
                        - STATUS_ON
                        - STATUS_OFF
                    type: string
                    format: enum
                remark:
                    type: string
        UpdateMenuResponse:
            type: object
            properties:
                menu:
                    $ref: '#/components/schemas/Menu'
...
        UploadFileRequest:
            type: object
            properties:
                file_name:
                    type: string
                file_directory:
                    type: string
                content:
                    type: string
                    format: bytes
        UploadFileResponse:
            type: object
            properties:
                file:
                    $ref: '#/components/schemas/File'
...